Remove override of IQueryable
