+++ DataTables.NetStandard/Extensions/QueryableExtensions.cs
@@ -8,7 +8,7 @@ using DataTables.NetStandard.Util;
 
 namespace DataTables.NetStandard.Extensions
 {
-    public static class QueryableExtensions
+    public static partial class QueryableExtensions
     {
         /// <summary>
         /// Builds a <see cref="PagedList{TEntity, TEntityViewModel}"/> from the given <see cref="IQueryable{T}"/>.
@@ -23,7 +23,7 @@ namespace DataTables.NetStandard.Extensions
             this IQueryable<TEntity> queryable,
             DataTablesRequest<TEntity, TEntityViewModel> request)
         {
-            return queryable.Apply(request).ToPagedList();
+            return new PagedList<TEntity, TEntityViewModel>(queryable.Apply(request), request);
         }
 
         /// <summary>
@@ -39,7 +39,7 @@ namespace DataTables.NetStandard.Extensions
             this IQueryable<TEntity> queryable,
             DataTablesRequest<TEntity, TEntityViewModel> request)
         {
-            return queryable.Apply(request).ToPagedListAsync();
+            return Task.Factory.StartNew<IPagedList<TEntityViewModel>>(() => new PagedList<TEntity, TEntityViewModel>(queryable.Apply(request), request));
         }
 
         /// <summary>
@@ -51,14 +51,14 @@ namespace DataTables.NetStandard.Extensions
         /// <typeparam name="TEntityViewModel">The type of the entity view model.</typeparam>
         /// <param name="queryable">The queryable.</param>
         /// <param name="request">The request.</param>
-        public static IDataTablesQueryable<TEntity, TEntityViewModel> Apply<TEntity, TEntityViewModel>(
+        public static IQueryable<TEntity> Apply<TEntity, TEntityViewModel>(
             this IQueryable<TEntity> queryable,
             DataTablesRequest<TEntity, TEntityViewModel> request)
         {
-            queryable = queryable.AsDataTablesQueryable(request)
-                .ApplyGlobalSearchFilter()
-                .ApplyColumnSearchFilter()
-                .ApplyOrder();
+            queryable = queryable
+                .ApplyGlobalSearchFilter(request)
+                .ApplyColumnSearchFilter(request)
+                .ApplyOrder(request);
 
             if (request.Log != null)
             {
@@ -76,22 +76,7 @@ namespace DataTables.NetStandard.Extensions
                 request.Log.Invoke(sb.ToString());
             }
 
-            return (IDataTablesQueryable<TEntity, TEntityViewModel>)queryable;
-        }
-
-        /// <summary>
-        /// Converts the given <see cref="IQueryable{T}"/> and the given <see cref="DataTablesRequest{TEntity, TEntityViewModel}"/>
-        /// to a new <see cref="DataTablesQueryable{TEntity, TEntityViewModel}"/>.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TEntityViewModel">The type of the entity view model.</typeparam>
-        /// <param name="queryable">The queryable.</param>
-        /// <param name="request">The request.</param>
-        public static IDataTablesQueryable<TEntity, TEntityViewModel> AsDataTablesQueryable<TEntity, TEntityViewModel>(
-            this IQueryable<TEntity> queryable,
-            DataTablesRequest<TEntity, TEntityViewModel> request)
-        {
-            return new DataTablesQueryable<TEntity, TEntityViewModel>(queryable, request);
+            return queryable;
         }
 
         /// <summary>
@@ -171,6 +156,197 @@ namespace DataTables.NetStandard.Extensions
 
                 _ => nameof(Queryable.OrderBy),
             };
+        }
+
+        /// <summary>
+        /// Applies the <see cref="DataTablesRequest{TEntity, TEntityViewModel}.GlobalSearchValue"/> of the given
+        /// <see cref="IQueryable{TEntity, TEntityViewModel}"/> to the query, if set.
+        /// To perform the global search, the method will add a search expression for each searchable column of the request.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <typeparam name="TEntityViewModel">The type of the entity view model.</typeparam>
+        /// <param name="queryable"></param>
+        /// <param name="request"></param>
+        internal static IQueryable<TEntity> ApplyGlobalSearchFilter<TEntity, TEntityViewModel>(
+            this IQueryable<TEntity> queryable,
+            DataTablesRequest<TEntity, TEntityViewModel> request)
+        {
+            var globalSearchValue = request.GlobalSearchValue;
+            if (!string.IsNullOrEmpty(globalSearchValue))
+            {
+                var columns = request.Columns.Where(c => c.IsSearchable);
+
+                if (columns.Any())
+                {
+                    Expression<Func<TEntity, bool>> predicate = null;
+
+                    foreach (var c in columns)
+                    {
+                        var globalSearchPredicateByProvider = c.GlobalSarchPredicateProvider != null
+                            ? c.GlobalSarchPredicateProvider(globalSearchValue)
+                            : null;
+
+                        var searchPredicateByProvider = globalSearchPredicateByProvider == null && c.SearchPredicateProvider != null
+                            ? c.SearchPredicateProvider(globalSearchValue)
+                            : null;
+
+                        Expression<Func<TEntity, bool>> expression;
+
+                        var searchPredicate = globalSearchPredicateByProvider
+                            ?? searchPredicateByProvider
+                            ?? c.GlobalSearchPredicate
+                            ?? c.SearchPredicate;
+
+                        if (searchPredicate != null)
+                        {
+                            var expr = searchPredicate;
+                            var entityParam = expr.Parameters.Single(p => p.Type == typeof(TEntity));
+                            var searchValueConstant = Expression.Constant(globalSearchValue, typeof(string));
+                            expression = (Expression<Func<TEntity, bool>>)Expression.Lambda(
+                                Expression.Invoke(expr, entityParam, searchValueConstant),
+                                entityParam);
+                        }
+                        else if (request.GlobalSearchRegex)
+                        {
+                            expression = ExpressionHelper.BuildRegexPredicate<TEntity>(c.PrivatePropertyName, globalSearchValue);
+                        }
+                        else
+                        {
+                            expression = ExpressionHelper.BuildStringContainsPredicate<TEntity>(
+                                c.PrivatePropertyName,
+                                globalSearchValue,
+                                c.SearchCaseInsensitive);
+                        }
+
+                        predicate = predicate == null
+                            ? expression
+                            : predicate.Or(expression);
+                    }
+
+                    if (predicate != null)
+                    {
+                        queryable = queryable.Where(predicate);
+                    }
+                }
+            }
+
+            return queryable;
+        }
+
+        /// <summary>
+        /// Applies the search filter for each of the searchable <see cref="DataTablesRequest{TEntity, TEntityViewModel}.Columns"/> 
+        /// where a search value is present.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <typeparam name="TEntityViewModel">The type of the entity view model.</typeparam>
+        /// <param name="queryable"></param>
+        /// <param name="request"></param>
+        internal static IQueryable<TEntity> ApplyColumnSearchFilter<TEntity, TEntityViewModel>(
+            this IQueryable<TEntity> queryable,
+            DataTablesRequest<TEntity, TEntityViewModel> request)
+        {
+            var columns = request.Columns
+                .Where(c => c.IsSearchable && !string.IsNullOrEmpty(c.SearchValue));
+
+            if (columns.Any())
+            {
+                Expression<Func<TEntity, bool>> predicate = null;
+
+                foreach (var c in columns)
+                {
+                    var columnSearchPredicateByProvider = c.ColumnSearchPredicateProvider != null
+                        ? c.ColumnSearchPredicateProvider(c.SearchValue)
+                        : null;
+
+                    var searchPredicateByProvider = columnSearchPredicateByProvider == null && c.SearchPredicateProvider != null
+                        ? c.SearchPredicateProvider(c.SearchValue)
+                        : null;
+
+                    Expression<Func<TEntity, bool>> expression;
+
+                    var searchPredicate = columnSearchPredicateByProvider
+                        ?? searchPredicateByProvider
+                        ?? c.ColumnSearchPredicate
+                        ?? c.SearchPredicate;
+
+                    if (searchPredicate != null)
+                    {
+                        var expr = searchPredicate;
+                        var entityParam = expr.Parameters.Single(p => p.Type == typeof(TEntity));
+                        var searchValueConstant = Expression.Constant(c.SearchValue, typeof(string));
+                        expression = (Expression<Func<TEntity, bool>>)Expression.Lambda(
+                            Expression.Invoke(expr, entityParam, searchValueConstant),
+                            entityParam);
+                    }
+                    else if (c.SearchRegex)
+                    {
+                        expression = ExpressionHelper.BuildRegexPredicate<TEntity>(c.PrivatePropertyName, c.SearchValue);
+                    }
+                    else
+                    {
+                        expression = ExpressionHelper.BuildStringContainsPredicate<TEntity>(
+                            c.PrivatePropertyName,
+                            c.SearchValue,
+                            c.SearchCaseInsensitive);
+                    }
+
+                    predicate = predicate == null
+                        ? expression
+                        : predicate.And(expression);
+                }
+
+                if (predicate != null)
+                {
+                    queryable = queryable.Where(predicate);
+                }
+            }
+
+            return queryable;
+        }
+
+        /// <summary>
+        /// Applies the requested order for each of the <see cref="DataTablesRequest{TEntity, TEntityViewModel}.Columns"/>.
+        /// </summary>
+        /// <typeparam name="TEntity">The type of the entity.</typeparam>
+        /// <typeparam name="TEntityViewModel">The type of the entity view model.</typeparam>
+        /// <param name="queryable"></param>
+        /// <param name="request"></param>
+        internal static IQueryable<TEntity> ApplyOrder<TEntity, TEntityViewModel>(
+            this IQueryable<TEntity> queryable,
+            DataTablesRequest<TEntity, TEntityViewModel> request)
+        {
+            var columns = request.Columns
+                .Where(c => c.IsOrderable && c.OrderingIndex >= 0)
+                .OrderBy(c => c.OrderingIndex);
+
+            bool alreadyOrdered = false;
+
+            foreach (var c in columns)
+            {
+                if (c.ColumnOrderingExpression != null)
+                {
+                    queryable = queryable.OrderBy(
+                        c.ColumnOrderingExpression,
+                        c.OrderingDirection,
+                        alreadyOrdered);
+                }
+                else
+                {
+                    var propertyName = c.ColumnOrderingProperty != null
+                        ? c.ColumnOrderingProperty.GetPropertyPath()
+                        : c.PrivatePropertyName;
+
+                    queryable = queryable.OrderBy(
+                        propertyName,
+                        c.OrderingDirection,
+                        c.OrderingCaseInsensitive,
+                        alreadyOrdered);
+                }
+
+                alreadyOrdered = true;
+            }
+
+            return queryable;
         }
     }
 }
